+++ packages/kit/src/core/adapt/builder.js
@@ -1,9 +1,8 @@
 import { existsSync, statSync, createReadStream, createWriteStream } from 'node:fs';
-import { join } from 'node:path/posix';
+import { extname, resolve } from 'node:path';
 import { pipeline } from 'node:stream';
 import { promisify } from 'node:util';
 import zlib from 'node:zlib';
-import glob from 'tiny-glob';
 import { copy, rimraf, mkdirp } from '../../utils/filesystem.js';
 import { generate_manifest } from '../generate_manifest/index.js';
 import { get_route_segments } from '../../utils/routing.js';
@@ -11,8 +10,10 @@ import { get_env } from '../../exports/vite/utils.js';
 import generate_fallback from '../postbuild/fallback.js';
 import { write } from '../sync/utils.js';
 import { load_error_page } from '../config/index.js';
+import { list_files } from '../utils.js';
 
 const pipe = promisify(pipeline);
+const extensions = ['.html', '.js', '.mjs', '.json', '.css', '.svg', '.xml', '.wasm'];
 
 /**
  * Creates the Builder which is passed to adapters for building the application.
@@ -84,15 +85,12 @@ export function create_builder({
 				return;
 			}
 
-			const files = await glob('**/*.{html,js,mjs,json,css,svg,xml,wasm}', {
-				cwd: directory,
-				dot: true,
-				absolute: true,
-				filesOnly: true
-			});
+			const files = list_files(directory, (file) => extensions.includes(extname(file))).map(
+				(file) => resolve(directory, file)
+			);
 
 			await Promise.all(
-				files.map((file) => Promise.all([compress_file(file, 'gz'), compress_file(file, 'br')]))
+				files.flatMap((file) => [compress_file(file, 'gz'), compress_file(file, 'br')])
 			);
 		},
 
@@ -161,7 +159,7 @@ export function create_builder({
 			return load_error_page(config);
 		},
 
-		generateManifest: ({ relativePath, routes: subset }) => {
+		generateManifest({ relativePath, routes: subset }) {
 			return generate_manifest({
 				build_data,
 				relative_path: relativePath,
@@ -188,23 +186,10 @@ export function create_builder({
 		},
 
 		writeClient(dest) {
-			const server_assets = copy(
-				`${config.kit.outDir}/output/server/${config.kit.appDir}/immutable/assets`,
-				join(dest, config.kit.appDir, 'immutable/assets')
-			).map((filename) => join(config.kit.appDir, 'immutable/assets', filename));
-			const client_assets = copy(`${config.kit.outDir}/output/client`, dest);
-			return Array.from(new Set([...server_assets, ...client_assets]));
+			return copy(`${config.kit.outDir}/output/client`, dest);
 		},
 
-		// @ts-expect-error
-		writePrerendered(dest, opts) {
-			// TODO remove for 1.0
-			if (opts?.fallback) {
-				throw new Error(
-					'The fallback option no longer exists — use builder.generateFallback(fallback) instead'
-				);
-			}
-
+		writePrerendered(dest) {
 			const source = `${config.kit.outDir}/output/prerendered`;
 			return [...copy(`${source}/pages`, dest), ...copy(`${source}/dependencies`, dest)];
 		},