+++ sites/kit.svelte.dev/src/routes/+page.svelte
@@ -1,13 +1,14 @@
 <script>
+	import { Footer, TrySection } from '@sveltejs/site-kit/home';
 	import { onMount } from 'svelte';
+	import Deployment from './home/Deployment.svelte';
 	import Features from './home/Features.svelte';
 	import Hero from './home/Hero.svelte';
-	import Showcase from './home/Showcase.svelte';
-	import Try from './home/Try.svelte';
-	import Deployment from './home/Deployment.svelte';
-	import Svelte from './home/Svelte.svelte';
 	import Intro from './home/Intro.svelte';
+	import Showcase from './home/Showcase.svelte';
+	import Svelte from './home/Svelte.svelte';
 	import schema_url from './schema.json?url';
+
 	import './home/common.css';
 
 	let schema;
@@ -48,41 +49,52 @@
 
 	<Hero />
 	<Intro />
-	<Try />
+	<TrySection />
 	<Svelte />
 	<Features />
 	<Deployment />
 	<Showcase />
 
-	<footer>
-		<p>
+	<Footer
+		links={{
+			resources: [
+				{
+					title: 'documentation',
+					href: '/docs'
+				},
+				{
+					title: 'tutorial',
+					href: 'https://learn.svelte.dev/tutorial/introducing-sveltekit'
+				},
+				{
+					title: 'blog',
+					href: 'https://svelte.dev/blog'
+				}
+			],
+			connect: [
+				{
+					title: 'github',
+					href: 'https://github.com/sveltejs/kit'
+				},
+				{
+					title: 'opencollective',
+					href: 'https://opencollective.com/svelte'
+				},
+				{
+					title: 'discord',
+					href: 'https://svelte.dev/chat'
+				},
+				{
+					title: 'twitter',
+					href: 'https://twitter.com/sveltejs'
+				}
+			]
+		}}
+	>
+		<span slot="license">
 			SvelteKit is <a target="_blank" rel="noreferrer" href="https://github.com/sveltejs/kit"
 				>free and open source software</a
 			> released under the MIT license.
-		</p>
-	</footer>
+		</span></Footer
+	>
 </div>
-
-<style>
-	footer {
-		padding: 1em var(--sk-page-padding-side);
-		text-align: center;
-		background: var(--sk-back-2);
-	}
-
-	footer p {
-		max-width: 20em;
-		margin: 0 auto;
-	}
-
-	footer p a {
-		color: inherit;
-		text-decoration: underline;
-	}
-
-	@media (min-width: 680px) {
-		footer p {
-			max-width: none;
-		}
-	}
-</style>